requirements.txt
httpx>=0.27.0
click>=8.0.0
pyyaml>=6.0.0
markitdown>=0.1.0
python-dotenv>=1.0.0
pytest-asyncio>=0.23.0
